# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Trançify SaaS multi-tenant scheduling API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: getMyAppointments
      tags: [appointments]
      summary: Get appointments for current tenant
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, cancelled, completed]
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get a single appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update an appointment (tenant)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentRequest"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /appointments/book:
    post:
      operationId: bookAppointment
      tags: [appointments]
      summary: Book a new appointment (client, public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookAppointmentRequest"
      responses:
        "201":
          description: Appointment booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          description: Slot unavailable or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /appointments/{id}/cost:
    patch:
      operationId: updateAppointmentCost
      tags: [appointments]
      summary: Update material cost for an appointment (tenant)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCostRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Appointment:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        serviceId:
          type: string
        serviceName:
          type: string
        clientName:
          type: string
        clientAge:
          type: integer
          nullable: true
        clientPhone:
          type: string
          nullable: true
        hairDescription:
          type: string
          nullable: true
        referencePhotos:
          type: array
          items:
            type: string
        paymentMethod:
          type: string
          enum: [pix, card, cash]
        braidSize:
          type: string
          enum: [mid_back, waist_butt]
        servicePrice:
          type: number
        materialCost:
          type: number
          nullable: true
        profit:
          type: number
          nullable: true
        date:
          type: string
          format: date
        time:
          type: string
          description: HH:MM
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        token:
          type: string
          description: Client editing token
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, tenantId, serviceId, serviceName, clientName, paymentMethod, braidSize, servicePrice, date, time, status, token, createdAt]
    BookAppointmentRequest:
      type: object
      properties:
        tenantId:
          type: string
        serviceId:
          type: string
        clientName:
          type: string
        clientAge:
          type: integer
        clientPhone:
          type: string
        hairDescription:
          type: string
        referencePhotos:
          type: array
          items:
            type: string
        paymentMethod:
          type: string
          enum: [pix, card, cash]
        braidSize:
          type: string
          enum: [mid_back, waist_butt]
        date:
          type: string
          format: date
        time:
          type: string
        notes:
          type: string
      required: [tenantId, serviceId, clientName, paymentMethod, braidSize, date, time]
    UpdateAppointmentRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        date:
          type: string
          format: date
        time:
          type: string
        notes:
          type: string
    UpdateCostRequest:
      type: object
      properties:
        materialCost:
          type: number
      required: [materialCost]
...